Pull-up resister and USB charging
Questions answered. Basing higher pull-up resistors for I2C to lower the
battery drain. The board is not user expandable to add more devices to the I2C bus.
The board needs to operate while USB is connected and charging the battery.
I anticipate users plugging in a USB powerbank to keep the watch going as they play games on it.