Recorder experiment added
Recorder is a utility to capture and create media archive files for playing on the Reflections entertainment platform. It records video and audio using WebRTC in a browser window and uploads it to a server over HTTPS. Recorder’s server side component processes the video into 240x240 pixel uncompressed mjpeg format, packs the video and audio into a TAR file, creates a Show manifest, and provides file transfer services to the Reflections device.