New scalable logger
I wrote a logger. It runs on an ESP32-S3, stores the log values on local storage temporarily, and sends the log values over Wifi to a service for storage. My goal is to have an unlimited number of log entries. I do not want to bog-down the ESP32 as it stores and uploads the logs. And I plan to have hundreds or thousands of devices logging concurrently.