Recorder project update
Cloud City are the server-side services to support the Reflections mobile entertainment experiences. Reflections Recorder records video and audio using WebRTC in a browser window, uploads it to a server over HTTPS, processes the video into 240x240 mjpeg movie and audio formats, packs the video and audio into a TAR file, and downloads the TAR to a Reflections board over Wifi to play.