Getting ready for D23
We've decided to show off the wrist watch at Disney D23 fan convention August 9-11, 2024. The next few weeks of commits are in support of the demonstration. We have these features planned:

1) Awake and deep sleep
2) Sending vibrations to other watches nearby - gestures (accel or TOF) and BLE communication
3) Mystic watch - gesture + answer text animation
4) Extra credit - Parallax cat or follow eyes - TOF follow
5) Extra credit - Set and Show Time

This commit changes the TOF sensor to recognize gestures from finger movements over the watch. It refines the Experience/Inveigle service. And enhances the GPS and BLE support.