Maker Faire 2024 Maxwell Demo
Frank Cohen exhbited the Reflections board and technology at the Maker Faire Bay Area 2024.
A foamcore board has a Reflections board, battery, and display to show a fortune telling
entertaining experience. Shake the board and a hopeful, benign, and negative answer
appears on the display. The flamboyant parrot knods in agreement for hopeful answers,
nods left and right for benign ansers, and bows his head for negative answers.